Allow compilation of the programs individually. Improve the installation procedure.
git-svn-id: http://gedanken.org.uk/svn/procmeter3/trunk@135 901458b3-6183-4e2b-843c-571e48fa976d

+++ Makefile
@@ -1,4 +1,4 @@
-# $Header: /home/amb/CVS/procmeter3/Makefile,v 1.12 2001-01-04 19:25:45 amb Exp $
+# $Header: /home/amb/CVS/procmeter3/Makefile,v 1.13 2001-01-05 19:49:53 amb Exp $
 #
 # ProcMeter - A system monitoring program for Linux - Version 3.3.
 #
@@ -6,7 +6,7 @@
 #
 # Written by Andrew M. Bishop
 #
-# This file Copyright 1994,95,96,97,98,99,2000 Andrew M. Bishop
+# This file Copyright 1994,95,96,97,98,99,2000,01 Andrew M. Bishop
 # It may be distributed under the GNU Public License, version 2, or
 # any higher version.  See section COPYING of the GNU Public license
 # for conditions under which this file may be redistributed.
@@ -15,7 +15,7 @@
 # Paths
 
 # INSTDIR - the default root directory for the package
-INSTDIR=/usr/local
+INSTDIR=/tmp/procmeter/usr/local
 
 # LIB_PATH - the root dir for library files
 LIB_PATH=$(INSTDIR)/lib/X11/ProcMeter3
@@ -46,16 +46,13 @@ OBJ=$(foreach f,$(SRC),$(addsuffix .o,$(basename $f)))
 
 ########
 
-all : obj \
-      procmeter3 gprocmeter3 \
-      procmeter.modules \
-      procmeterrc.install
+all : procmeter3 gprocmeter3
 
 ########
 
-procmeter3  : procmeter.xaw
+procmeter3  : obj procmeter.xaw procmeter.modules procmeterrc.install
 
-gprocmeter3 : procmeter.gtk
+gprocmeter3 : obj procmeter.gtk procmeter.modules procmeterrc.install
 
 ########
 
@@ -110,19 +107,26 @@ distclean : clean
 
 .PHONY : install
 
-install : all
+install :
+	@[ -f procmeter3 ] || [ -f gprocmeter3 ] || (echo "*** Run 'make all' or 'make procmeter3' or 'make gprocmeter3' first." ; exit 1)
 	install -d $(LIB_PATH)
+	install -d $(MOD_PATH)
 	$(MAKE) -C modules install MOD_PATH=$(MOD_PATH) LIB_PATH=$(LIB_PATH)
 	install -d $(INSTDIR)/bin
-	install -m 755 procmeter3 $(INSTDIR)/bin
-	install -m 755 gprocmeter3 $(INSTDIR)/bin
+	[ ! -f procmeter3 ] || install -m 755 procmeter3 $(INSTDIR)/bin
+	@[ -f procmeter3 ] || (echo "" ; echo "*** The procmeter3 program has not been installed (it does not exist)." ; echo "")
+	[ ! -f gprocmeter3 ] || install -m 755 gprocmeter3 $(INSTDIR)/bin
+	@[ -f gprocmeter3 ] || (echo "" ; echo "*** The gprocmeter3 program has not been installed (it does not exist)." ; echo "")
 	install -d $(INSTDIR)/man/man1
 	install -d $(INSTDIR)/man/man5
-	install -m 644 procmeter.1 $(INSTDIR)/man/man1/procmeter3.1
-	install -m 644 procmeter.1 $(INSTDIR)/man/man1/gprocmeter3.1
+	[ ! -f procmeter3 ] || install -m 644 procmeter3.1 $(INSTDIR)/man/man1
+	[ ! -f gprocmeter3 ] || install -m 644 gprocmeter3.1 $(INSTDIR)/man/man1
 	install -m 644 procmeterrc.5 $(INSTDIR)/man/man5/procmeterrc.5
 	[ ! -f $(LIB_PATH)/.procmeterrc ] || mv $(LIB_PATH)/.procmeterrc $(LIB_PATH)/procmeterrc
-	[ ! -f $(RC_PATH) ] || mv $(RC_PATH) $(RC_PATH).bak
-	install -m 644 procmeterrc.install $(RC_PATH)
+	[ ! -f $(RC_PATH) ] || for n in 0 1 2 3 4 5 6 7 8 9; do \
+				  [ ! -f $(RC_PATH) -o -f $(RC_PATH).bak.$$n ] || mv $(RC_PATH) $(RC_PATH).bak.$$n ; \
+				  done
+	@[ ! -f $(RC_PATH) ] || (echo "" ; echo "*** The $(RC_PATH) file has not been installed (it already exists)." ; echo "")
+	[ -f $(RC_PATH) ] || install -m 644 procmeterrc.install $(RC_PATH)
 	install -d $(LIB_PATH)/include
 	install -m 644 procmeter.h $(LIB_PATH)/include
